Let Transforms handle complex structures and fix the tfm_y issue #245
To train a mask rcnn using Fastai we need to return  Mask and Bounding
boxes for each  sample, this currently cannot be achieved using fastai.

The proposed changes:
- Fix the above issue, as it gives user ability to define arbitrary
complex structures in x and y.

- Allow users to use transform_on_side even if  tfm_y is PXIEL, MASK or
COORD.

- Moreover the introduction of .determ() will let us remove the Thread
local storage from set_state in a long run, probably speeding up the
transformations, and making them less dependant on side effects.

The propose change works as follows:

We remove assumption that we have only x and y returned from dataset.
Instead we allow user to specify a custom function that is responsible
for applying a deterministic transformation to the structure of x and
y.

See TestTransforms.test_it_is_possible_to_transform_mask_n_coords for a
concrete example.